+++ lib/puppet/type/aem_user.rb
@@ -77,4 +77,13 @@ Puppet::Type.newtype(:aem_user)do
     desc 'Group name'
   end
 
+  newparam :force do
+    desc 'Set to true to force user creation, if the user already exists then it will be deleted before recreated, default to false'
+    validate do |value|
+      if value == ''
+        value = false
+      end
+    end
+  end
+
 end